Move bauble_test to be exported
